Added functionality, fixed crashes in certain situations
Added functionality: 1. Clicking outside of the keyboard/appbar turns the searchbar off.
2. Going to the buildings page turns the searchbar off.
Known issues/expected behavior to add: 3. Replicate (2) for the about page. 4. Make the back button hide the searchbar. 5. Make the map not resize when keyboard appears. 6. Reset the home page to the default look when user changes route.